fixed datatype conversion warnings

+++ SDLGame2D/SDLGame2D/cSprite.cpp
@@ -14,7 +14,7 @@ cSprite::cSprite(SDL_Texture* texture, float x, float y, int w, int h, float nom
 
 void cSprite::DrawSprite(SDL_Renderer* renderer)
 {
-	SDL_Rect tempRect = { mPosX, mPosY, mWidth, mHeight };
+	SDL_Rect tempRect = { (int)mPosX, (int)mPosY, (int)mWidth, (int)mHeight };
 	SDL_RenderCopy(renderer, mTexture, NULL, &tempRect);
 }
